# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/workspaces/{workspaceId}/budget": {
      "get": {
        "operationId": "PlanningController_getBudget_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One payload per DEC-011; spend derived from postings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Budget status: planned vs actual for the current period",
        "tags": [
          "Planning"
        ]
      },
      "put": {
        "operationId": "PlanningController_upsertBudget_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertBudgetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create or replace the budget and all of its lines",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/budget/{id}": {
      "delete": {
        "operationId": "PlanningController_deleteBudget_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a budget",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/goals": {
      "get": {
        "operationId": "PlanningController_getGoals_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalsOverviewResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Savings goals and debt summary, progress read from the ledger",
        "tags": [
          "Planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_createGoal_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalProgressResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a savings goal",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/goals/{id}": {
      "patch": {
        "operationId": "PlanningController_updateGoal_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalProgressResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Update a savings goal (progress is not writable)",
        "tags": [
          "Planning"
        ]
      },
      "delete": {
        "operationId": "PlanningController_deleteGoal_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a savings goal",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/debts": {
      "put": {
        "operationId": "PlanningController_upsertDebt_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDebtDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Attach or replace repayment terms on a liability account",
        "tags": [
          "Planning"
        ]
      },
      "get": {
        "operationId": "PlanningController_listDebts_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtOverviewResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "List debt terms with outstanding balances derived from the ledger",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/debts/{accountId}": {
      "delete": {
        "operationId": "PlanningController_deleteDebt_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Remove debt terms without deleting the liability account or balance",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/calendar": {
      "get": {
        "operationId": "PlanningController_getCalendar_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarOverviewResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Upcoming, due and overdue events",
        "tags": [
          "Planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_createEvent_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalendarEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a bill, expected income or reminder",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/calendar/{id}": {
      "patch": {
        "operationId": "PlanningController_updateEvent_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCalendarEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Mark paid or skipped, or edit an event",
        "tags": [
          "Planning"
        ]
      },
      "delete": {
        "operationId": "PlanningController_deleteEvent_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a calendar event",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/recurring": {
      "get": {
        "operationId": "PlanningController_listRules_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecurringRuleResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "List recurring rules",
        "tags": [
          "Planning"
        ]
      },
      "post": {
        "operationId": "PlanningController_createRule_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecurringRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Create a recurring rule (never auto-posts — §9.4)",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/recurring/{id}": {
      "patch": {
        "operationId": "PlanningController_updateRule_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecurringRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringRuleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Pause or resume a recurring rule",
        "tags": [
          "Planning"
        ]
      },
      "delete": {
        "operationId": "PlanningController_deleteRule_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Delete a recurring rule",
        "tags": [
          "Planning"
        ]
      }
    },
    "/v1/workspaces/{workspaceId}/reports/categories": {
      "get": {
        "operationId": "PlanningController_getReport_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive local date",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Exclusive local date",
            "schema": {
              "example": "2026-09-01",
              "type": "string"
            }
          },
          {
            "name": "granularity",
            "required": false,
            "in": "query",
            "schema": {
              "default": "month",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryReportResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Category breakdown and 6-month trend (§11.3 contract)",
        "tags": [
          "Planning"
        ]
      }
    },
...
  },
  "info": {
    "title": "NoorixFin API",
    "description": "Personal and household finance management API. All financial amounts are minor-unit decimal strings. All timestamps are ISO 8601 UTC. Every failure body matches the `ApiErrorBody` schema; branch on its `code`, never on `message`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "BudgetStatusResponseDto": {
        "type": "object",
        "properties": {
          "visible": {
            "type": "boolean"
          },
          "has_budget": {
            "type": "boolean"
          },
          "budget_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "cadence": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "WEEKLY"
            ]
          },
          "rollover": {
            "type": "boolean"
          },
          "period_start": {
            "type": "string"
          },
          "period_end": {
            "type": "string"
          },
          "planned_total": {
            "type": "number"
          },
          "spent_total": {
            "type": "number"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetLineResponseDto"
            }
          },
          "timezone": {
            "type": "string"
          },
          "generated_at": {
            "type": "string"
          }
        },
        "required": [
          "visible"
        ]
      },
...
      "DeletedResponseDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "DebtOverviewResponseDto": {
        "type": "object",
        "properties": {
          "debts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DebtResponseDto"
            }
          },
          "total_debt_minor": {
            "type": "number"
          }
        },
        "required": [
          "debts",
          "total_debt_minor"
        ]
      },
...
    }
  }
}